requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.1.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.12.3
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-appconf==1.0.3
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-tabbed-admin==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.5.6
oauthlib==2.1.0
openapi-codec==1.3.2
pep8==1.7.1
pilkit==2.0
Pillow==5.3.0
pycodestyle==2.4.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
zope.interface==4.6.0
